Feature/landing page (#35)
* initial working pages

* API populated dropdown update

* working draft

* working geojson features with tooltip

* remove unused modules

* working multipage app

* red volcano icon

* Flake8 style fixes

* remove test script

* Gracefull handle vrrc-API response errors

* Add missing unrest_table.csv

* move config.ini file to project root

* remove debug print statement

* move config.ini file

* missing config in dockerfile
