+++ app/Data/unrest_table.csv
@@ -1,0 +1,11 @@
+Site,Latest SAR Image Date,Unrest
+Meager,None,False
+Garibaldi,None,False
+Cayley,None,False
+Nazko,None,False
+Lava_Fork,None,False
+Tseax,None,False
+Edziza_North,None,False
+Edziza_South,None,False
+Hoodoo,None,False
+Edgecumbe,None,True
