+++ requirements.txt
@@ -38,3 +38,5 @@ tzdata==2023.3
 urllib3==1.26.15
 Werkzeug==2.2.3
 zipp==3.15.0
+requests
+pyyaml
